+++ src/onegov.form/onegov/form/upgrade.py
@@ -2,3 +2,10 @@
 upgraded on the server. See :class:`onegov.core.upgrade.upgrade_task`.
 
 """
+from onegov.core.upgrade import upgrade_task
+
+
+@upgrade_task('Enable external form submissions')
+def enable_external_form_submissions(context):
+
+    context.operations.alter_column('submissions', 'name', nullable=True)